Fix for Linux
The sun_len field appears to not be POSIX compatible and only exists
in BSD. It also doesn't _appear_ to be necessary based on [this
document](https://www.ibm.com/docs/en/i/7.1?topic=family-af-unix-address).

_Note: The sun_len field is provided only for BSD 4.4 compatibility. It
is not necessary to use this field even for BSD 4.4/ UNIX 98
compatibility. The field is ignored on input addresses._